backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
requests
beautifulsoup4
# ↓↓↓ この行が重要です！「0.5.0」以降の新しいバージョンを直接指定します。
google-generativeai==0.5.4